Fix memory leak, and can't connect bug
Memory leak happened when everyone left a lobby where there are bot/local players.
Can't connect happened when you created an online lobby, then left the lobby, and when you try to create a new lobby, because the "player.lobby" was not cleared, you couldn't create the lobby (game assumed you were already in a lobby).